// wheretogo-backend/wheretogo-backend/src/generated/typegraphql/resolvers/inputs/UserWhereInput.ts
export class UserWhereInput {
  @TypeGraphQL.Field(_type => [UserWhereInput], {
    nullable: true
  })
  AND?: UserWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [UserWhereInput], {
    nullable: true
  })
  OR?: UserWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [UserWhereInput], {
    nullable: true
  })
  NOT?: UserWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => IntFilter, {
    nullable: true
  })
  id?: IntFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  name?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  nickname?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  email?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  password?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => EnumUserRoleFilter, {
    nullable: true
  })
  role?: EnumUserRoleFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  provider?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  avatar?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  bio?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeFilter, {
    nullable: true
  })
  createdAt?: DateTimeFilter | undefined;

  @TypeGraphQL.Field(_type => FavoriteListRelationFilter, {
    nullable: true
  })
  favorites?: FavoriteListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => ReviewListRelationFilter, {
    nullable: true
  })
  reviews?: ReviewListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => BookingListRelationFilter, {
    nullable: true
  })
  bookings?: BookingListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => FollowerListRelationFilter, {
    nullable: true
  })
  followers?: FollowerListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => FollowerListRelationFilter, {
    nullable: true
  })
  following?: FollowerListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => NotificationListRelationFilter, {
    nullable: true
  })
  notifications?: NotificationListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => DeviceTokenListRelationFilter, {
    nullable: true
  })
  deviceTokens?: DeviceTokenListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => TripListRelationFilter, {
    nullable: true
  })
  tripsOwned?: TripListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => TripCollaboratorListRelationFilter, {
    nullable: true
  })
  tripsShared?: TripCollaboratorListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => ChatMessageListRelationFilter, {
    nullable: true
  })
  chatMessages?: ChatMessageListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => ChatParticipantListRelationFilter, {
    nullable: true
  })
  chatParticipants?: ChatParticipantListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => ChatMessageReadListRelationFilter, {
    nullable: true
  })
  messageReads?: ChatMessageReadListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => RefreshTokenListRelationFilter, {
    nullable: true
  })
  refreshTokens?: RefreshTokenListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => ScheduleLocationInCalendarListRelationFilter, {
    nullable: true
  })
  scheduleLocationInCalendars?: ScheduleLocationInCalendarListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PostListRelationFilter, {
    nullable: true
  })
  posts?: PostListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PostLikeListRelationFilter, {
    nullable: true
  })
  postLikes?: PostLikeListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PostCommentListRelationFilter, {
    nullable: true
  })
  postComments?: PostCommentListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => CommentLikeListRelationFilter, {
    nullable: true
  })
  commentLikes?: CommentLikeListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => SavedPostListRelationFilter, {
    nullable: true
  })
  savedPosts?: SavedPostListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PostCollectionListRelationFilter, {
    nullable: true
  })
  collections?: PostCollectionListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PostShareListRelationFilter, {
    nullable: true
  })
  postShares?: PostShareListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PostReportListRelationFilter, {
    nullable: true
  })
  postReports?: PostReportListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => UploadSessionListRelationFilter, {
    nullable: true
  })
  uploadSessions?: UploadSessionListRelationFilter | undefined;
}